requirements.txt
# Audit Hybride - Dépendances Python
# Installation: pip install -r requirements.txt

# HTTP Requests
requests>=2.28.0

# Variables d'environnement
python-dotenv>=1.0.0

# Désactivation warnings SSL
urllib3>=1.26.0
